+++ src/Ui/Bundle/EventBundle/Controller/Meeting/UpdateMeetingAction.php
@@ -6,16 +6,16 @@ use Proximum\Vimeet\Application\Adapter\AuthorizationCheckerAdapterInterface;
 use Proximum\Vimeet\Application\Adapter\CommandBusInterface;
 use Proximum\Vimeet\Application\Adapter\QueryBusInterface;
 use Proximum\Vimeet\Application\Adapter\TranslatorInterface;
-use Proximum\Vimeet\Application\Command\Meeting\Event\Move;
-use Proximum\Vimeet\Application\Exception\Meeting\MoveMeetingException;
+use Proximum\Vimeet\Application\Command\Meeting\Event\Update;
+use Proximum\Vimeet\Application\Exception\Meeting\UpdateMeetingException;
 use Proximum\Vimeet\Application\Query\MeetingSlot\GetAvailableSlotsQuery;
 use Proximum\Vimeet\Application\Query\MeetingSlot\GetAvailableSlotsView;
 use Proximum\Vimeet\Domain\Event\GetTimezoneHelper;
-use Proximum\Vimeet\Domain\Meeting\CanMoveMeeting;
+use Proximum\Vimeet\Domain\Meeting\CanUpdateMeeting;
 use Proximum\Vimeet\Domain\Model\Meeting;
 use Proximum\Vimeet\Domain\Model\Participant;
 use Proximum\Vimeet\Domain\Model\Sheet;
-use Proximum\Vimeet\Ui\Bundle\EventBundle\Form\Type\Meeting\MoveMeetingType;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\Form\Type\Meeting\UpdateMeetingType;
 use Proximum\Vimeet\Ui\Bundle\EventBundle\Security\SheetVoter;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\Request;
@@ -24,10 +24,10 @@ use Symfony\Component\HttpFoundation\Session\Flash\FlashBagInterface;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Templating\EngineInterface;
 
-class MoveMeetingAction
+class UpdateMeetingAction
 {
-    /** @var CanMoveMeeting */
-    private $canMoveMeeting;
+    /** @var CanUpdateMeeting */
+    private $canUpdateMeeting;
 
     /** @var FormFactoryInterface $formFactory */
     private $formFactory;
@@ -55,7 +55,7 @@ class MoveMeetingAction
 
     public function __construct(
         AuthorizationCheckerAdapterInterface $authorizationCheckerAdapter,
-        CanMoveMeeting $canMoveMeeting,
+        CanUpdateMeeting $canUpdateMeeting,
         FormFactoryInterface $formFactory,
         EngineInterface $engine,
         CommandBusInterface $commandBus,
@@ -65,7 +65,7 @@ class MoveMeetingAction
         TranslatorInterface $translator
     ) {
         $this->authorizationCheckerAdapter = $authorizationCheckerAdapter;
-        $this->canMoveMeeting = $canMoveMeeting;
+        $this->canUpdateMeeting = $canUpdateMeeting;
         $this->formFactory = $formFactory;
         $this->engine = $engine;
         $this->commandBus = $commandBus;
@@ -78,18 +78,19 @@ class MoveMeetingAction
     public function __invoke(Request $request, Participant $participant, Sheet $sheet, Meeting $meeting): Response
     {
         if (!$this->authorizationCheckerAdapter->isGranted(SheetVoter::EDIT, $sheet)
-            || false === $this->canMoveMeeting->isSatisfiedBy($sheet)
+            || false === $this->canUpdateMeeting->isSatisfiedBy($sheet)
             || !$meeting->hasSheet($sheet)
         ) {
             throw new AccessDeniedException();
         }
 
         /** @var GetAvailableSlotsView $availableSlotsView */
-        $availableSlotsView = $this->queryBus->handle(new GetAvailableSlotsQuery($meeting, $meeting->isVisio()));
+        $availableSlotsView = $this->queryBus->handle(new GetAvailableSlotsQuery($meeting, $meeting->isVisio(), $sheet));
         $timezone = $this->getTimezoneHelper->getTimezoneByEventAndParticipant($sheet->getEvent(), $participant);
 
-        $move = new Move($sheet, $meeting);
-        $form = $this->formFactory->create(MoveMeetingType::class, $move, [
+        $update = new Update($sheet, $meeting, $meeting->getParticipants($sheet), $meeting->getSlot());
+        $form = $this->formFactory->create(UpdateMeetingType::class, $update, [
+            'participants' => $sheet->getParticipantsArray(),
             'availableSlots' => $availableSlotsView->availableSlots,
             'timezone' => $timezone,
             'locale' => $sheet->getEvent()->getAvailableLocale($request->getLocale()),
@@ -98,14 +99,14 @@ class MoveMeetingAction
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             try {
-                $this->commandBus->handle($move);
-            } catch (MoveMeetingException $exception) {
+                $this->commandBus->handle($update);
+            } catch (UpdateMeetingException $exception) {
                 return new Response($exception->getMessage(), Response::HTTP_BAD_REQUEST);
             }
 
             $this->flashBag->set(
                 'success',
-                $this->translator->trans('agenda.meeting.move.success')
+                $this->translator->trans('agenda.meeting.update.success')
             );
 
             return new Response(null, Response::HTTP_NO_CONTENT);
@@ -118,8 +119,9 @@ class MoveMeetingAction
         }
 
         return new Response(
-            $this->engine->render('@Event/Meeting/move-meeting-slot-form.html.twig', [
+            $this->engine->render('@Event/Meeting/update-meeting-form.html.twig', [
                 'form' => $form->createView(),
+                'currentSheetAvailableSlots' => $availableSlotsView->currentSheetAvailableSlotIds,
             ])
         );
     }